Create rickroll
How to Use:
Save this script with a .dolphin file extension.
Transfer the script to your Flipper Zero device.
Set your Flipper Zero to Keyboard mode.
Connect the Flipper Zero to a target computer.
Run the script from the Flipper Zero.
This script simulates the pressing of keys to open a web browser and navigate to the Rick Astley "Never Gonna Give You Up" video on YouTube. Make sure to use this script responsibly and only on devices you own or have permission to use.

If you have any further questions or need assistance with something else, feel free to ask!